refactor: extract app action dispatch
Move App::process_action into its own dispatch module so App state and helpers are easier to navigate without changing behavior.
